Trivia quiz (very early alpha vesion)
Please note that the only function that woks is the trivia function. You can enter any number into the level select prompt and it will go to that line *10 and do 5 questions. the odd lines are the questions and even lines are the answer for the question above it. Found an issue? Report it.